Replace long with uint64_t for hostid.

+++ dmtcp/src/uniquepid.h
@@ -38,7 +38,7 @@ namespace dmtcp
     static dmtcp::UniquePid& ThisProcess(bool disableJTrace = false);
     UniquePid();
 
-    UniquePid ( const long& host, const pid_t& pd, const uint64_t& tm,
+    UniquePid ( const uint32_t& host, const pid_t& pd, const uint64_t& tm,
                 const int& gen = 0 ) {
       _hostid = host;
       _pid = pd;
@@ -54,7 +54,7 @@ namespace dmtcp
     }
 
     UniquePid(const char *str);
-    long hostid() const { return _hostid; }
+    uint64_t hostid() const { return _hostid; }
     pid_t pid() const { return _pid; }
     int generation() const { return _generation; }
     uint64_t time() const { return _time; }
